requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
openai==1.40.0
pydantic==2.8.2
httpx==0.27.2
